add documentation generation and README rendering functionality
