+++ .pre-commit-config.yaml
@@ -33,7 +33,7 @@ repos:
       - id: poetry-check
       - id: poetry-lock
   - repo: https://github.com/google/yamlfmt
-    rev: v0.13.0
+    rev: v0.14.0
     hooks:
       - id: yamlfmt
   - repo: https://github.com/PyCQA/flake8
@@ -41,3 +41,10 @@ repos:
     hooks:
       - id: flake8
         args: ["--max-line-length", "88"]
+  - repo: local
+    hooks:
+      - id: readme
+        name: rendering readme
+        entry: poetry run python docs/readme.py
+        language: system
+        files: docs/README\.md\.jinja2$
