+++ README.md
@@ -22,6 +22,8 @@ $ pip install poetry-plugin-hook
 
 or with `poetry`
 
+> Especially on [Windows](https://python-poetry.org/docs/cli/#self-add), self commands that update or remove packages may be problematic.
+
 ```shell
 $ poetry self add poetry-plugin-hook
 ```
@@ -32,11 +34,38 @@ Wrapper for `poetry show -o -T` command.
 
 Exit code represents the number of outdated packages.
 
-```bash
-$ poetry hook latest && echo exit-code: $?
+```shell
+$ poetry hook latest --help
 
-All top-level dependencies are up-to-date.
-exit-code: 0
+  Description:
+    Check if all top-level dependencies are up-to-date.
+
+  Usage:
+    hook latest [options] [--] [<package>]
+
+  Arguments:
+    package                    The package to inspect
+
+  Options:
+        --without=WITHOUT      The dependency groups to ignore. (multiple values allowed)
+        --with=WITH            The optional dependency groups to include. (multiple values allowed)
+        --only=ONLY            The only dependency groups to include. (multiple values allowed)
+    -l, --latest               Show the latest version. (option is always True)
+    -o, --outdated             Show the latest version but only for packages that are outdated. (option is always True)
+    -T, --top-level            Show only top-level dependencies. (option is always True)
+    -h, --help                 Display help for the given command. When no command is given display help for the list command.
+    -q, --quiet                Do not output any message.
+    -V, --version              Display this application version.
+        --ansi                 Force ANSI output.
+        --no-ansi              Disable ANSI output.
+    -n, --no-interaction       Do not ask any interactive question.
+        --no-plugins           Disables plugins.
+        --no-cache             Disables Poetry source caches.
+    -C, --directory=DIRECTORY  The working directory for the Poetry command (defaults to the current working directory).
+    -v|vv|vvv, --verbose       Increase the verbosity of messages: 1 for normal output, 2 for more verbose output and 3 for debug.
+
+  Help:
+    poetry hook latest [options]
 ```
 
 ## hook sync
@@ -45,11 +74,40 @@ Wrapper for `poetry install --sync` command.
 
 With `--exit` option, the command returns the corresponding value as exit code. With it's default `--exit=any` the sum of *installs*, *updates* and *removals* is returned.
 
-```bash
-$ poetry hook sync --dry-run && echo exit-code: $?
+```shell
+$ poetry hook sync --help
 
-No dependencies to install or update.
-exit-code: 0
+  Description:
+    Synchronize the environment with the locked packages and the specified groups.
+
+  Usage:
+    hook sync [options]
+
+  Options:
+        --exit=EXIT            Specify the value to return as exitcode. choices=['any', 'installs', 'updates', 'removals'] [default: "any"]
+        --without=WITHOUT      The dependency groups to ignore. (multiple values allowed)
+        --with=WITH            The optional dependency groups to include. (multiple values allowed)
+        --only=ONLY            The only dependency groups to include. (multiple values allowed)
+        --sync                 Synchronize the environment with the locked packages and the specified groups. (option is always True)
+        --no-root              Do not install the root package (the current project).
+        --no-directory         Do not install any directory path dependencies; useful to install dependencies without source code, e.g. for caching of Docker layers)
+        --dry-run              Output the operations but do not execute anything (implicitly enables --verbose).
+    -E, --extras=EXTRAS        Extra sets of dependencies to install. (multiple values allowed)
+        --all-extras           Install all extra dependencies.
+        --only-root            Exclude all dependencies.
+    -h, --help                 Display help for the given command. When no command is given display help for the list command.
+    -q, --quiet                Do not output any message.
+    -V, --version              Display this application version.
+        --ansi                 Force ANSI output.
+        --no-ansi              Disable ANSI output.
+    -n, --no-interaction       Do not ask any interactive question.
+        --no-plugins           Disables plugins.
+        --no-cache             Disables Poetry source caches.
+    -C, --directory=DIRECTORY  The working directory for the Poetry command (defaults to the current working directory).
+    -v|vv|vvv, --verbose       Increase the verbosity of messages: 1 for normal output, 2 for more verbose output and 3 for debug.
+
+  Help:
+    poetry hook sync [options]
 ```
 
 ## pre-commit-config
