+++ docs/__init__.py
@@ -1,0 +1,7 @@
+from .documentation import main as documentation
+from .readme import main as readme
+
+__all__ = [
+    "readme",
+    "documentation",
+]
