+++ docs/__main__.py
@@ -1,95 +1,15 @@
-import json
-import warnings
+from . import documentation, readme
 
 
-class WarningMessageEncoder(json.JSONEncoder):
-    def default(self, obj):
-
-        if isinstance(obj, warnings.WarningMessage):
-            return {
-                k: v
-                for k, v in obj.__dict__.items()
-                if v is not None and not k.startswith("_")
-            }
-
-        if isinstance(obj, type):
-            return obj.__name__
-
-        if isinstance(obj, Exception):
-            return str(obj).strip().split("\n")
-
-        return super().default(obj)
-
-
-def main() -> int:
+def main():
     """
-    Create a static html documentation for the project in the 'public' directory.
-
-    Requires follwing PyPi packages:
-    - 'pdoc'
-    - 'pathlibutil'
+    Render the README.md file and create a static html documentation for the project
+    in the 'public' directory.
 
     Returns non-zero on failure.
     """
 
-    try:
-        from pathlibutil import Path
-        from pdoc import pdoc, render
-
-        with Path(__file__).parent as cwd:
-            print(f"\nrunning docs in {cwd=}...")
-
-            documentation = cwd / "public"
-
-            try:
-                print(f"cleaning output directory {documentation.size()}.")
-                documentation.delete(recursive=True, missing_ok=True)
-            except FileNotFoundError:
-                pass
-
-            documentation.mkdir(parents=True)
-
-            config = {
-                "template_directory": cwd / "dark-mode",
-                "show_source": False,
-                "search": False,
-                "docformat": "google",
-            }
-
-            modules = [
-                "poetry_plugin_hook",
-                "poetry_plugin_hook.redirect",
-                "poetry",
-                "cleo",
-            ]
-
-            with warnings.catch_warnings(record=True) as messages:
-                print(f"rendering {modules=}.")
-
-                render.configure(**config)
-                pdoc(*modules, output_directory=documentation)
-
-                log = documentation / "warnings.json"
-
-                with log.open("w", encoding="utf-8") as f:
-                    json.dump(
-                        messages,
-                        f,
-                        cls=WarningMessageEncoder,
-                    )
-
-                if messages:
-                    print(f"{log=} generated with {len(messages)} warnings.")
-    except ModuleNotFoundError as e:
-        print(f"Creation failed, due missing dependency!\n\tpip install {e.name}")
-        return 2
-    except Exception as e:
-        print(f"{documentation=} creation failed!\n\t{e}")
-        return 1
-
-    print(f"{documentation=} generated successfully.\n")
-
-    return 0
+    return readme() or documentation()
 
 
 if __name__ == "__main__":
