+++ pyproject.toml
@@ -61,4 +61,7 @@ addopts = [
 ]
 
 [tool.coverage.run]
-omit = [ "*/tests/*" ]
+omit = [
+  "*/tests/*",
+  "*/docs/*",
+]
